app/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
django
djangorestframework
dj-rest-auth
djangorestframework-simplejwt